==> health/Health Care Services/Physicians/Primary Care/Service Access Scores/legacy/2.catchment_scores_primcare_va.Rmd <==
---
title: "catchment_scores"
author: "--"
date: "2/4/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
# Libraries
```{r}
library(catchment)
library(sf)
library(leaflet)
library(dplyr)
```
# Start new
```{r}
 rm(list=ls())
```
# Working directory
```{r}
setwd("~/VDH/Floating_Catchment_Areas/va/4. primcare_va_model/")
```
# Read data
```{r}
population <- read.csv("population.csv")
provider <- read.csv("provider_20.csv")
rownames(provider) <- provider$ID
traveltimes <- read.csv("traveltimes_trimmed_20.csv", row.names = 1)
population$prov_cnt[is.na(population$prov_cnt)] <- 0
```
# Prepare map
```{r}

blockgroup_shapes <-   st_transform( tigris::block_groups(state = c("VA"), class = "sf", year=2020), 4326)

rownames(blockgroup_shapes) <- blockgroup_shapes$GEOID
blockgroup_shapes <- blockgroup_shapes[as.character(population$GEOID),]
##new
shape_counties <- tigris::counties(state = c("VA"), class = "sf")
#format for counties
shape_counties <- st_transform(shape_counties, 4326 )
#create map
map <- leaflet(blockgroup_shapes, options = leafletOptions(attributionControl = FALSE)) %>%
  addTiles() %>%
  addScaleBar("bottomleft") %>%
  addMapPane("lines", zIndex = 410) %>%
  addMapPane("points", zIndex = 411) %>%
  addPolygons(
    fillColor = colorNumeric("RdYlBu", population$population)(population$population),
    fillOpacity = 1, stroke = FALSE, group = "Population", label = population$pediatric_pop
  ) %>%
  hideGroup("Population") %>%
  addLayersControl(
    position = "topleft", overlayGroups = c( "Population", "Access")
  ) %>%
  addCircles(
    data = provider, color = "blue", lng = ~lon, lat = ~lat, weight = 3,
    label = ~ paste0("ID: ", ID),  options = pathOptions(pane = "points")
  )
```
# Estimate catchment areas
```{r}
#2sfca

population$providers_2sfca <- catchment_ratio(
  consumers = population, 
  providers = provider, 
  cost = traveltimes,
  weight = 30, 
  consumers_id = "GEOID", 
  consumers_value = "population",
  consumers_location = c("X", "Y"), 
  providers_id = "ID",
  providers_value = "doctors", 
  providers_location = c("lon", "lat"),
  verbose = FALSE)*1000

population_plot <- population %>% mutate(state=substr(GEOID,1,2)) %>% filter(state==51)

pal <- colorBin("RdYlBu", population_plot$providers_2sfca)

plot_2sfca <- map %>%
  addControl("Primary Care Doctors Per 1,000 People (2-Step Floating Catchment Area)", "topright") %>%
  addLegend("bottomright", pal, population_plot$providers_2sfca, opacity = .7) %>%
  addPolygons(
    fillColor = pal(population_plot$providers_2sfca), fillOpacity = .7, weight = 1, color = "#000",
    highlightOptions = highlightOptions(color = "#fff"), group = "Access",
    label = paste0(
      "GEOID: ", population_plot$GEOID, ", Population: ", population_plot$population,
      ", Per 1k People: ", round(population_plot$providers_2sfca, 4), ", In Region: ",
      round(population_plot$population * population_plot$providers_2sfca / 1000, 4)
    )
  ) %>%
  addPolylines(data = shape_counties, color = "black", opacity = 1, weight = 2.5) 

## enhanced 2-step

step_weights <- list(c(30, .22), c(20, .68), c(10, 1))

population$providers_e2sfca <- catchment_ratio(
  consumers = population, 
  providers = provider, 
  cost = traveltimes,
  weight = step_weights, 
  consumers_id = "GEOID", 
  consumers_value = "population",
  consumers_location = c("X", "Y"), 
  providers_id = "ID",
  providers_value = "doctors", 
  providers_location = c("lon", "lat"),
  verbose = FALSE)*1000

#df for plot
population_plot <- population %>% mutate(state=substr(GEOID,1,2)) %>% filter(state==51)

pal_e2sfca <- colorBin("RdYlBu", population_plot$providers_e2sfca)
 
plot_e2sfca <- map %>%
  addControl("Primary Care Doctors Per 1,000 People (enhanced 2-Step Floating Catchment Area)", "topright") %>%
  addLegend("bottomright", pal_e2sfca, population_plot$providers_e2sfca, opacity = .7) %>%
  addPolygons(
    fillColor = pal_e2sfca(population_plot$providers_e2sfca), fillOpacity = .7, weight = 1, color = "#000",
    highlightOptions = highlightOptions(color = "#fff"), group = "Access",
    label = paste0(
      "GEOID: ", population_plot$GEOID, ", Population: ", population_plot$population,
      ", Per 1k People: ", round(population_plot$providers_e2sfca, 4), ", In Region: ",
      round(population_plot$population * population_plot$providers_e2sfca / 1000, 4)
    )
  ) %>% addPolylines(data = shape_counties, color = "black", opacity = 1, weight = 2.5) 


## 3-step

population$providers_3sfca <- catchment_ratio(
  consumers = population, 
  providers = provider, 
  cost = traveltimes,
  weight = step_weights, 
  normalize_weight = TRUE,
  consumers_id = "GEOID", 
  consumers_value = "population",
  consumers_location = c("X", "Y"), 
  providers_id = "ID",
  providers_value = "doctors", 
  providers_location = c("lon", "lat"),
  verbose = FALSE)*1000

#df for plot
population_plot <- population %>% mutate(state=substr(GEOID,1,2)) %>% filter(state==51)

pal_3sfca <- colorBin("RdYlBu", population_plot$providers_3sfca)

plot_3sfca <- map %>%
  addControl("Primary Care Doctors Per 1,000 People (3-Step Floating Catchment Area)", "topright") %>%
  addLegend("bottomright", pal_3sfca, population_plot$providers_3sfca, opacity = .7) %>%
  addPolygons(
    fillColor = pal_3sfca(population_plot$providers_3sfca), fillOpacity = .7, weight = 1, color = "#000",
    highlightOptions = highlightOptions(color = "#fff"), group = "Access",
    label = paste0(
      "GEOID: ", population_plot$GEOID, ", Population: ", population_plot$population,
      ", Per 1k People: ", round(population_plot$providers_3sfca, 4), ", In Region: ",
      round(population_plot$population * population_plot$providers_3sfca / 1000, 4)
    )
  ) %>% addPolylines(data = shape_counties, color = "black", opacity = 1, weight = 2.5) 

```

# Plot Catchment Areas 
```{r}
plot_2sfca
plot_e2sfca
plot_3sfca
```


# Plot distribution of Catchment Areas
```{r}
library(ggplot2)
library(hrbrthemes)
library(dplyr)
library(tidyr)
library(viridis)

Data_1 <- data.frame( Observation = population_plot$providers_2sfca , fca = '2sfca')
Data_2 <- data.frame( Observation = population_plot$providers_e2sfca , fca = 'e2sfca')
Data_3 <- data.frame( Observation = population_plot$providers_3sfca, fca = '3sfca' )

CombinedData <- rbind (Data_1, Data_2, Data_3)

densities_fca <- ggplot(data=CombinedData, aes(x=Observation, group=fca, fill=fca)) +
    geom_density(adjust=1.5, alpha=.4) +
    theme_ipsum() +
  scale_fill_viridis(discrete=TRUE) +
  labs(title = 'Distribution of Floating Catchment Areas - Virginia - Pediatrics')

densities_fca
```
### Aggregation of Catchment Areas Information
# Reshape wide-long
```{r}
#DEFINE relevant population and names
dmv_bg_2021_acccess_scores_wide <- population_plot   %>% dplyr::select( geoid="GEOID", 
                                                            primcare_pop_cnt ="population",
                                                            primcare_cnt = "prov_cnt",
                                                            primcare_near_10_mean = "near_10_mean",
                                                            primcare_near_10_median = "near_10_median",
                                                            primcare_2sfca = "providers_2sfca" ,
                                                            primcare_e2sfca = "providers_e2sfca", 
                                                            primcare_3sfca = "providers_3sfca" )
```

# Long format - block group
```{r}
dmv_bg_2021_acccess_scores_bg <- tidyr::gather(dmv_bg_2021_acccess_scores_wide, measure, value, 
                                            primcare_pop_cnt,
                                            primcare_cnt,
                                            primcare_near_10_mean,
                                            primcare_near_10_median,
                                            primcare_2sfca, 
                                            primcare_e2sfca, 
                                            primcare_3sfca)
#define year
dmv_bg_2021_acccess_scores_bg$year <- 2021
#define measure type
dmv_bg_2021_acccess_scores_bg$measure_type <- as.factor(ifelse(dmv_bg_2021_acccess_scores_bg$measure=="primcare_2sfca" |
                                                            dmv_bg_2021_acccess_scores_bg$measure=="primcare_e2sfca" |
                                                              dmv_bg_2021_acccess_scores_bg$measure=="primcare_3sfca", 'index','count') )
```

# Include name: geographic name for geographic unit: block group
```{r}
#a) connect
con <- RPostgreSQL::dbConnect(drv = RPostgreSQL::PostgreSQL(),
                              dbname = "sdad",
                              host = "postgis1",
                              port = 5432,
                              user = Sys.getenv(x = "DB_USR"),
                              password = Sys.getenv(x = "DB_PWD"))

#b) query
name_geo <- sf::st_read(con, 
  query= "
SELECT  *
FROM dc_geographies.ncr_cttrbg_tiger_2010_2020_geo_names "
)

#join region_name + region_type
dmv_bg_2021_acccess_scores_bg$geoid <- as.character(dmv_bg_2021_acccess_scores_bg$geoid)
dmv_bg_2021_acccess_scores_bg <- dmv_bg_2021_acccess_scores_bg %>% left_join(name_geo, by="geoid")
#order
dmv_bg_2021_acccess_scores_bg <- dmv_bg_2021_acccess_scores_bg %>% dplyr::select("geoid","region_type","region_name", "measure", "value", "year", "measure_type")
```

# Aggregate information for tracts
```{r}
#generate id tract 
#the relevant file is WIDE
dmv_bg_2021_acccess_scores_wide_aggreg <- dmv_bg_2021_acccess_scores_wide  %>% mutate(tract_geoid = substr(geoid,1,11))
#add region name - tract
# dmv_bg_2021_acccess_scores_wide_aggreg <- dmv_bg_2021_acccess_scores_wide_aggreg %>% left_join(name_geo, by= c("tract_geoid"= "geoid") )

## measures tract
dmv_bg_2021_acccess_scores_tract_wide <- dmv_bg_2021_acccess_scores_wide_aggreg %>% group_by(tract_geoid) %>% dplyr::summarise(
  primcare_pop = sum(primcare_pop_cnt, na.rm = TRUE), 
  primcare_cnt = sum(primcare_cnt, na.rm = TRUE), 
  primcare_near_10_mean = weighted.mean(primcare_near_10_mean, primcare_pop_cnt, na.rm = TRUE), 
  primcare_near_10_median = weighted.mean(primcare_near_10_median, primcare_pop_cnt, na.rm = TRUE), 
  primcare_2sfca = weighted.mean(primcare_2sfca, primcare_pop_cnt, na.rm = TRUE), 
  primcare_e2sfca = weighted.mean(primcare_e2sfca, primcare_pop_cnt, na.rm = TRUE),
  primcare_3sfca = weighted.mean(primcare_3sfca, primcare_pop_cnt, na.rm = TRUE),
  )

               
```

# Long format - tract 
```{r}
dmv_bg_2021_acccess_scores_tract <- tidyr::gather(dmv_bg_2021_acccess_scores_tract_wide, measure, value, 
                                            primcare_pop,
                                            primcare_cnt,
                                            primcare_near_10_mean,
                                            primcare_near_10_median,
                                            primcare_2sfca, 
                                            primcare_e2sfca, 
                                            primcare_3sfca)
#define year
dmv_bg_2021_acccess_scores_tract$year <- 2021
#define measure type
dmv_bg_2021_acccess_scores_tract$measure_type <- as.factor(ifelse(dmv_bg_2021_acccess_scores_tract$measure=="primcare_2sfca" |
                                                            dmv_bg_2021_acccess_scores_tract$measure=="primcare_e2sfca" |
                                                              dmv_bg_2021_acccess_scores_tract$measure=="primcare_3sfca", 'index','count') )
```

# Include name: geographic name for geographic unit: tract
```{r}
#join region_name + region_type
dmv_bg_2021_acccess_scores_tract$tract_geoid <- as.character(dmv_bg_2021_acccess_scores_tract$tract_geoid)
dmv_bg_2021_acccess_scores_tract <- dmv_bg_2021_acccess_scores_tract %>% left_join(name_geo, by= c("tract_geoid" = "geoid") )
#order
dmv_bg_2021_acccess_scores_tract <- dmv_bg_2021_acccess_scores_tract %>% dplyr::select(geoid="tract_geoid","region_type","region_name", "measure", "value", "year", "measure_type")
```

# Aggregate information for county
```{r}
#generate id tract 
#the relevant file is WIDE
dmv_bg_2021_acccess_scores_wide_aggreg <- dmv_bg_2021_acccess_scores_wide  %>% mutate(county_geoid = substr(geoid,1,5))
#add region name - tract
# dmv_bg_2021_acccess_scores_wide_aggreg <- dmv_bg_2021_acccess_scores_wide_aggreg %>% left_join(name_geo, by= c("tract_geoid"= "geoid") )

## measures tract
dmv_bg_2021_acccess_scores_county_wide <- dmv_bg_2021_acccess_scores_wide_aggreg %>% group_by(county_geoid) %>% summarise(
  primcare_pop = sum(primcare_pop_cnt, na.rm = TRUE), 
  primcare_cnt = sum(primcare_cnt, na.rm = TRUE), 
  primcare_near_10_mean = weighted.mean(primcare_near_10_mean, primcare_pop_cnt, na.rm = TRUE), 
  primcare_near_10_median = weighted.mean(primcare_near_10_median, primcare_pop_cnt, na.rm = TRUE), 
  primcare_2sfca = weighted.mean(primcare_2sfca, primcare_pop_cnt, na.rm = TRUE), 
  primcare_e2sfca = weighted.mean(primcare_e2sfca, primcare_pop_cnt, na.rm = TRUE),
  primcare_3sfca = weighted.mean(primcare_3sfca, primcare_pop_cnt, na.rm = TRUE),
  )

               
```

# Long format - county
```{r}
dmv_bg_2021_acccess_scores_county <- tidyr::gather(dmv_bg_2021_acccess_scores_county_wide, measure, value, 
                                            primcare_pop,
                                            primcare_cnt,
                                            primcare_near_10_mean,
                                            primcare_near_10_median,
                                            primcare_2sfca, 
                                            primcare_e2sfca, 
                                            primcare_3sfca)
#define year
dmv_bg_2021_acccess_scores_county$year <- 2021
#define measure type
dmv_bg_2021_acccess_scores_county$measure_type <- as.factor(ifelse(dmv_bg_2021_acccess_scores_county$measure=="primcare_2sfca" |
                                                            dmv_bg_2021_acccess_scores_county$measure=="primcare_e2sfca" |
                                                              dmv_bg_2021_acccess_scores_county$measure=="primcare_3sfca", 'index','count') )
```

# Include name: geographic name for geographic unit: county
```{r}
#join region_name + region_type
dmv_bg_2021_acccess_scores_county$tract_geoid <- as.character(dmv_bg_2021_acccess_scores_county$county_geoid)
dmv_bg_2021_acccess_scores_county <- dmv_bg_2021_acccess_scores_county %>% left_join(name_geo, by= c("county_geoid" = "geoid") )
#order
dmv_bg_2021_acccess_scores_county <- dmv_bg_2021_acccess_scores_county %>% dplyr::select(geoid="county_geoid","region_type","region_name", "measure", "value", "year", "measure_type")
```

# Health Districts categories
#####################################################
```{r}
health_district <- read.csv("~/spatial_access_health/Locality-to-HD-to-HPR.csv")
names(health_district)[2] <- "county_geoid"
health_district$county_geoid <- as.character(health_district$county_geoid )


```

# Aggregate information for Health district
```{r}

#join name of HD
dmv_bg_2021_acccess_scores_wide_aggreg <- dmv_bg_2021_acccess_scores_wide_aggreg %>% left_join(health_district, by="county_geoid")

#generate id tract 
#the relevant file is WIDE
#dmv_bg_2021_acccess_scores_wide_aggreg <- dmv_bg_2021_acccess_scores_wide  %>% mutate(tract_geoid = substr(geoid,1,11))
#add region name - tract
# dmv_bg_2021_acccess_scores_wide_aggreg <- dmv_bg_2021_acccess_scores_wide_aggreg %>% left_join(name_geo, by= c("tract_geoid"= "geoid") )

## measures tract
dmv_bg_2021_acccess_scores_hd_wide <- dmv_bg_2021_acccess_scores_wide_aggreg %>% group_by(HealthDistrict) %>% dplyr::summarise(
  primcare_pop = sum(primcare_pop_cnt, na.rm = TRUE), 
  primcare_cnt = sum(primcare_cnt, na.rm = TRUE), 
  primcare_near_10_mean = weighted.mean(primcare_near_10_mean, primcare_pop_cnt, na.rm = TRUE), 
  primcare_near_10_median = weighted.mean(primcare_near_10_median, primcare_pop_cnt, na.rm = TRUE), 
  primcare_2sfca = weighted.mean(primcare_2sfca, primcare_pop_cnt, na.rm = TRUE), 
  primcare_e2sfca = weighted.mean(primcare_e2sfca, primcare_pop_cnt, na.rm = TRUE),
  primcare_3sfca = weighted.mean(primcare_3sfca, primcare_pop_cnt, na.rm = TRUE)
  )

               
```

# Long format - Health district
```{r}
dmv_bg_2021_acccess_scores_hd <- tidyr::gather(dmv_bg_2021_acccess_scores_hd_wide, measure, value, 
                                            primcare_pop,
                                            primcare_cnt,
                                            primcare_near_10_mean,
                                            primcare_near_10_median,
                                            primcare_2sfca, 
                                            primcare_e2sfca, 
                                            primcare_3sfca)
#define year
dmv_bg_2021_acccess_scores_hd$year <- 2021
#define measure type
dmv_bg_2021_acccess_scores_hd$measure_type <- as.factor(ifelse(dmv_bg_2021_acccess_scores_hd$measure=="primcare_2sfca" |
                                                            dmv_bg_2021_acccess_scores_hd$measure=="primcare_e2sfca" |
                                                              dmv_bg_2021_acccess_scores_hd$measure=="primcare_3sfca", 'index','count') )
```

# Include name: geographic name for geographic unit: Health district
```{r}

con <- RPostgreSQL::dbConnect(drv = RPostgreSQL::PostgreSQL(),
                              dbname = "sdad",
                              host = "postgis1",
                              port = 5432,
                              user = Sys.getenv(x = "DB_USR"),
                              password = Sys.getenv(x = "DB_PWD"))
#b)query
hd_names  <- sf::st_read(
  con, 
  query= "
    SELECT *
 FROM dc_geographies.va_hd_vdh_2021_health_district_geo_names"
)

#create region_type and region_name: join 
dmv_bg_2021_acccess_scores_hd <- dmv_bg_2021_acccess_scores_hd %>% left_join(hd_names, by = c("HealthDistrict"="region_name") )
#order
dmv_bg_2021_acccess_scores_hd <- dmv_bg_2021_acccess_scores_hd %>% dplyr::select("geoid","region_type", region_name = "HealthDistrict", "measure", "value", "year", "measure_type")

```

# Combine measures of Block groups + Tracts + County, and Health Districts
```{r}
va_bg_2021_access_scores_primcare <- rbind(dmv_bg_2021_acccess_scores_bg, 
                                    dmv_bg_2021_acccess_scores_tract,
                                    dmv_bg_2021_acccess_scores_county, 
                                    dmv_bg_2021_acccess_scores_hd)

# Specify names for counts
va_bg_2021_access_scores_primcare$measure <- ifelse(va_bg_2021_access_scores_primcare$measure=='primcare_pop', 'primcare_pop_cnt' , va_bg_2021_access_scores_primcare$measure)
```


# Save information in Working directory and PgAdmin
```{r}

# write.csv(va_bg_2021_access_scores_primcare, "va_hdcttrbg_webmd_2021_access_scores_primcare.csv", row.names = FALSE )
readr::write_csv(va_bg_2021_access_scores_primcare, xzfile("va_hdcttrbgca_webmd_2021_access_scores_primcare.csv.xz", compression = 9))

```
